+++ lesson8/src/main/java/dev/tutushkin/lesson8/data/movies/remote/MoviesApi.kt
@@ -23,7 +23,7 @@ interface MoviesApi {
 
     @GET("movie/{movie_id}")
     suspend fun getMovieDetails(
-        @Path("movie_id") movieId: Long,
+        @Path("movie_id") movieId: Int,
         @Query("api_key") apiKey: String
     ): MovieDetailsResponse
