+++ lesson8/src/main/java/dev/tutushkin/lesson8/domain/movies/MoviesRepository.kt
@@ -12,5 +12,5 @@ interface MoviesRepository {
 
     suspend fun getNowPlaying(apiKey: String): Result<List<MovieList>>
 
-    suspend fun getMovieDetails(movieId: Long, apiKey: String): Result<MovieDetails>
+    suspend fun getMovieDetails(movieId: Int, apiKey: String): Result<MovieDetails>
 }
