+++ lesson8/src/main/java/dev/tutushkin/lesson8/presentation/moviedetails/viewmodel/MovieDetailsViewModelFactory.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.ExperimentalSerializationApi
 @ExperimentalSerializationApi
 class MovieDetailsViewModelFactory(
     private val repository: MoviesRepository,
-    private val id: Long
+    private val id: Int
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
